refactor: move sidebar tab item settings support
